Refactor: Update import paths and add new route in AppRouter
