+++ src/components/Sidebar.jsx
@@ -1,6 +1,6 @@
 import { NavLink, Outlet } from "react-router-dom";
 import Switch from "./Switch";
-import SoulJourneyLogo from "./sidebar/SoulJourneyLogo";
+import SoulJourneyLogo from "./SoulJourneyLogo";
 import SidebarListItems from "./sidebar/SidebarListItems";
 import useAuthCall from "../hooks/useAuthCall";
 import LanguageSelector from "./LanguageSelector";
@@ -14,13 +14,13 @@ const Sidebar = () => {
       <div className=" w-dvw h-dvh grid grid-cols-7">
         {/* SideBar */}
         <div className="col-span-1 bg-white">
-          <div className="p-2 h-full w-full flex flex-col bg-offWhite dark:bg-navy border-r border-r-gray-200">
+          <div className="xs:pb-2 sm:p-2 h-full w-full flex flex-col bg-offWhite dark:bg-navy border-r border-r-gray-200">
             {/* Logo */}
             <NavLink to="#">
-              <div className="flex flex-col justify-center lg:justify-start items-center gap-1 pb-2 px-0 md:px-2 lg:px-4 cursor-pointer dark:bg-offWhite dark:rounded-md dark:shadow-sm dark:shadow-offWhite-light">
+              <div className="flex flex-col justify-center lg:justify-start items-center gap-1 pb-2 px-0 xs:px-0 md:px-2 lg:px-4 cursor-pointer dark:bg-offWhite dark:xs:rounded-none dark:sm:rounded-md dark:shadow-sm dark:shadow-offWhite-light">
                 <SoulJourneyLogo/>
                 <Switch/>
-                <div className="bg-navy/90 rounded-full m-2 shadow-lg shadow-navy-light hover:bg-navy/70 hover:shadow-6xl hover:shadow-navy-dark">
+                <div className="w-14 flex justify-center bg-navy/90 rounded-xl m-4 px-0 shadow-lg shadow-navy-light hover:bg-navy/70 hover:shadow-6xl hover:shadow-navy-dark">
                   <LanguageSelector/>
                 </div>
               </div>
