+++ src/components/sidebar/SidebarListItems.jsx
@@ -10,7 +10,7 @@ const menu = [
     name: "Dashboard",
   },
   {
-    to: NavLinkto(""),
+    to: NavLinkto("appointment"),
     icon: icon("appointment"),
     name: "Appointment",
   },
@@ -74,7 +74,7 @@ const SidebarListItems = () => {
             </NavLink>
           ))}
         </div>
-        <div className="flex flex-col  space-y-1 mx-1 lg:mt-1 ">
+        <div className="flex flex-col  space-y-1 mx-1 mb-2 lg:mt-1 ">
           <NavLink
             className={({isActive})=> `group flex flex-row items-center  justify-center lg:justify-start rounded-md h-12 focus:outline-none pr-1.5 lg:pr-6 font-semibold text-navy-dark hover:text-seaGreen-dark cursor-pointer ${isActive ? "bg-seaGreen-dark/20 text-seaGreen-dark dark:bg-offWhite-dark/20 dark:text-offWhite-dark" : ""}`}
             to="/profile/account"
@@ -102,7 +102,7 @@ const SidebarListItems = () => {
           </NavLink>
         </div>
       </div>
-      <hr className="hidden lg:block"/>
+      <hr className="xs:hidden lg:block "/>
       {/* Sidebar footer */}
       {/* <div className="px-1" onClick={()=>logout()}>
         <div className="group flex flex-row items-center  justify-center lg:justify-start rounded-md h-12 focus:outline-none pr-1.5 lg:pr-6 font-semibold text-navy-dark dark:text-offWhite-dark hover:text-pink-light cursor-pointer  ">
