+++ src/pages/Appointment.jsx
@@ -1,0 +1,9 @@
+import React from 'react'
+
+const Appointment = () => {
+  return (
+    <div>Appointment</div>
+  )
+}
+
+export default Appointment
